First commit.
First commit; first attempt at a spider, using urllib2 and some regular
expressions.  Generally unoptimized; generally not error-resistant; etc.
Works for most things I've tried, but not streamlined for the user.

Included is an example of a function you can call with it: the get_title
function will take the title off of the web page and return it to you,
for each page in your crawled list.

TODO: Everything.